Add class for Image handle - ImageGenerator
